requirements.txt
django
django-suit

django-compressor
sorl-thumbnail
-e git+https://github.com/chriscauley/Django-Crop-Override-Field.git#egg=django-crop-override-field
pillow

python-social-auth
graphene-django
